Add hooks to Scheduler nodes for generating device-specific debug strings
